src/third_party/blink/web_tests/virtual/web-components-v0-disabled/fast/dom/shadow/shadow-boundary-events-expected.txt

CONSOLE ERROR: line 87: Uncaught TypeError: element.createShadowRoot is not a function
Tests to ensure that shadow DOM boundary is not crossed during event propagation. Can only run within DRT.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

FAIL successfullyParsed should be true. Was false.

TEST COMPLETE
